Refactor to remove GeometryEditorToolbarSettings.
